+++ changes.md
@@ -9,3 +9,17 @@
 - added property `shippingKey` to type `StagedOrderSetShippingMethodTaxRateAction`
 </details>
 
+
+<details>
+<summary>Added Resource(s)</summary>
+
+- added resource `/{projectKey}/as-associate/{associateId}/in-business-unit/key={businessUnitKey}/carts/replicate`
+</details>
+
+
+<details>
+<summary>Added Method(s)</summary>
+
+- added method `$apiRoot->withProjectKey()->asAssociate()->withAssociateIdValue()->inBusinessUnitKeyWithBusinessUnitKeyValue()->carts()->replicate()->post()`
+</details>
+